requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
transformers>=4.20.0
datasets>=2.0.0
nltk>=3.7
spacy>=3.4.0
wordcloud>=1.9.0
tensorboard>=2.9.0
wandb>=0.13.0
pyyaml>=6.0
tqdm>=4.64.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0



